arrumando NaN

+++ backend/src/controls/PartidaController.ts
@@ -81,11 +81,15 @@ static async create(req: Request, res: Response) {
 
   static async getById(req: Request, res: Response) {
     try {
-      const { id } = req.params;
+      const id = Number(req.params.id);
+      if (isNaN(id) || id <= 0) {
+        return res.status(400).json({ error: "ID inválido" });
+      }
+
       const repo = AppDataSource.getRepository(Partida);
 
       const partida = await repo.findOne({
-        where: { id: Number(id) },
+        where: { id },
         relations: ["local", "tipoPartida"],
       });
 
@@ -102,10 +106,14 @@ static async create(req: Request, res: Response) {
 
   static async update(req: Request, res: Response) {
     try {
-      const { id } = req.params;
-      const partidaRepo = AppDataSource.getRepository(Partida);
+      const id = Number(req.params.id);
+      if (isNaN(id) || id <= 0) {
+        return res.status(400).json({ error: "ID inválido" });
+      }
 
-      const partida = await partidaRepo.findOneBy({ id: Number(id) });
+      const partidaRepo = AppDataSource.getRepository(Partida);
+      const partida = await partidaRepo.findOneBy({ id });
+
       if (!partida) {
         return res.status(404).json({ error: "Partida não encontrada" });
       }
@@ -125,9 +133,14 @@ static async create(req: Request, res: Response) {
     }
   }
 
+
   static async delete(req: Request, res: Response) {
     try {
-      const { id } = req.params;
+      const id = Number(req.params.id);
+      if (isNaN(id) || id <= 0) {
+        return res.status(400).json({ error: "ID inválido" });
+      }
+
       const partidaRepo = AppDataSource.getRepository(Partida);
 
       const resultado = await partidaRepo.delete(id);
@@ -172,18 +185,18 @@ static async create(req: Request, res: Response) {
   }
 
   static async getPublicas(req: Request, res: Response) {
-  try {
-    const repo = AppDataSource.getRepository(Partida);
+    try {
+      const repo = AppDataSource.getRepository(Partida);
 
-    const publicas = await repo.find({
-      where: { tipo: TipoEnum.PUBLICO } // agora funciona suave
-    });
-    console.log("Partidas públicas encontradas:", publicas);
+      const publicas = await repo.find({
+        where: { tipo: TipoEnum.PUBLICO } // agora funciona suave
+      });
+      console.log("Partidas públicas encontradas:", publicas);
 
-    return res.json(publicas);
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ error: "Erro ao buscar partidas públicas" });
+      return res.json(publicas);
+    } catch (error) {
+      console.error(error);
+      return res.status(500).json({ error: "Erro ao buscar partidas públicas" });
+    }
   }
-}
 }
